Oppgave table header filter dropdowns: Fetch own kodeverk only when necessary
